+++ Editor/Behaviours/Ghost.h
@@ -6,7 +6,7 @@ class Ghost : public Behaviour
 public:
     DEFINE_BEHAVIOUR(Ghost);
 
-    void Update(Scene* Scene, float DeltaTime) override;
+    void Update(Scene* Scene, double DeltaTime) override;
 
     void SetTarget(Model* Target) { this->Target = Target; }
     float GhostSpeed = 4.0f;
