+++ Editor/Behaviours/Spin.cpp
@@ -6,8 +6,8 @@ Spin::Spin()
 {
 }
 
-void Spin::Update(Scene* Scene, float DeltaTime)
+void Spin::Update(Scene* Scene, double DeltaTime)
 {
     float Speed = 0.5f * 3.1415f;
-    m_Model->GetTransform().Rotate(Quaternion(Vec3f(0.0f, 0.0f, 1.0f), Speed * DeltaTime));
+    m_Model->GetTransform().Rotate(Quaternion(Vec3f(0.0f, 0.0f, 1.0f), Speed * (float)DeltaTime));
 }
