+++ Editor/Behaviours/Wobble.h
@@ -11,7 +11,7 @@ public:
 
     Wobble();
 
-    virtual void Update(Scene* Scene, float DeltaTime);
+    virtual void Update(Scene* Scene, double DeltaTime);
 
 private:
